pyproject.toml
[project]
name = "opaiui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Shawn T O'Neil", email = "oneilsh@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=0.4.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "upstash-redis>=1.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
